Rename package `output` to `tools`
