+++ moveit/CMakeLists.txt
@@ -5,7 +5,7 @@ find_package(Boost REQUIRED)
 
 find_package(catkin REQUIRED COMPONENTS
   moveit_benchmark_suite_core
-  moveit_benchmark_suite_output
+  moveit_benchmark_suite_tools
 )
 
 catkin_package(
@@ -15,7 +15,7 @@ catkin_package(
     include
   CATKIN_DEPENDS
     moveit_benchmark_suite_core
-    moveit_benchmark_suite_output
+    moveit_benchmark_suite_tools
 )
 
 set(CMAKE_CXX_STANDARD 14)
