+++ moveit/package.xml
@@ -10,7 +10,7 @@
 
 	<depend>boost</depend>
 	<depend>moveit_benchmark_suite_core</depend>
-	<depend>moveit_benchmark_suite_output</depend>
+	<depend>moveit_benchmark_suite_tools</depend>
 
 	<export>
 		<moveit_benchmark_suite benchmarks="${prefix}/benchmarks"/>
