+++ tools/config/gnuplot.yaml
