+++ tools/config/plots.yaml
