+++ tools/launch/plot_dataset.launch
@@ -8,9 +8,9 @@
   <arg unless="$(arg debug)" name="launch_prefix" value="" />
   <arg     if="$(arg debug)" name="launch_prefix" value="gdb --args" />
 
-  <node name="plot_dataset" pkg="moveit_benchmark_suite_output" type="plot_dataset" respawn="false" clear_params="true" required="true" output="screen" launch-prefix="$(arg launch_prefix)">
+  <node name="plot_dataset" pkg="moveit_benchmark_suite_tools" type="plot_dataset" respawn="false" clear_params="true" required="true" output="screen" launch-prefix="$(arg launch_prefix)">
 
-    <param name="config_file" value="$(find moveit_benchmark_suite_output)/config/gnuplot.yaml"/>
+    <param name="config_file" value="$(find moveit_benchmark_suite_tools)/config/gnuplot.yaml"/>
     <rosparam param="input_files" subst_value="True">$(arg input_files)</rosparam>
 
   </node>
