+++ tools/package.xml
@@ -1,5 +1,5 @@
 <package format="2">
-	<name>moveit_benchmark_suite_output</name>
+	<name>moveit_benchmark_suite_tools</name>
 	<version>0.0.0</version>
 	<description>Store, plot and visualize components of the MoveIt Benchmark Suite</description>
