+++ tools/src/generate_plots.cpp
@@ -1,16 +1,17 @@
 
 #include <ros/ros.h>
-#include <moveit_benchmark_suite/profiler.h>
-#include <moveit_benchmark_suite/output/gnuplot.h>
+#include <moveit_benchmark_suite/tools/gnuplot.h>
+#include <moveit_benchmark_suite/tools/htmlplot.h>
+#include <moveit_benchmark_suite/benchmark.h>
 
 using namespace moveit_benchmark_suite;
-using namespace moveit_benchmark_suite::output;
+using namespace moveit_benchmark_suite::tools;
 
 constexpr char INPUT_PARAMETER[] = "input_files";
 
 int main(int argc, char** argv)
 {
-  ros::init(argc, argv, "plot_dataset");
+  ros::init(argc, argv, "generate_plots");
   ros::AsyncSpinner spinner(1);
   spinner.start();
 
@@ -22,14 +23,24 @@ int main(int argc, char** argv)
   pnh.getParam(INPUT_PARAMETER, dataset_files);
   pnh.getParam(CONFIG_PARAMETER, config_file);
 
-  // Plot dataset
+  // Generate GNUPlot script
   GNUPlotDataset gnuplot;
 
   gnuplot.initializeFromYAML(config_file);
   gnuplot.plot(dataset_files);
 
-  ROS_WARN("Press Ctl-C to terminate GNUPlot");
-  ros::waitForShutdown();
+  // Add GNUPlot instances into HTML
+  HTMLPlot html;
+  auto names = gnuplot.getInstanceNames();
+
+  for (const auto& name : names)
+  {
+    std::string output;
+    gnuplot.getInstanceOutput(name, output);
+    html.writeline(output);
+  }
+
+  html.dump();
 
   return 0;
 }
