+++ tools/src/plot_dataset.cpp
@@ -1,17 +1,16 @@
 
 #include <ros/ros.h>
-#include <moveit_benchmark_suite/output/gnuplot.h>
-#include <moveit_benchmark_suite/output/htmlplot.h>
-#include <moveit_benchmark_suite/benchmark.h>
+#include <moveit_benchmark_suite/profiler.h>
+#include <moveit_benchmark_suite/tools/gnuplot.h>
 
 using namespace moveit_benchmark_suite;
-using namespace moveit_benchmark_suite::output;
+using namespace moveit_benchmark_suite::tools;
 
 constexpr char INPUT_PARAMETER[] = "input_files";
 
 int main(int argc, char** argv)
 {
-  ros::init(argc, argv, "generate_plots");
+  ros::init(argc, argv, "plot_dataset");
   ros::AsyncSpinner spinner(1);
   spinner.start();
 
@@ -23,24 +22,14 @@ int main(int argc, char** argv)
   pnh.getParam(INPUT_PARAMETER, dataset_files);
   pnh.getParam(CONFIG_PARAMETER, config_file);
 
-  // Generate GNUPlot script
+  // Plot dataset
   GNUPlotDataset gnuplot;
 
   gnuplot.initializeFromYAML(config_file);
   gnuplot.plot(dataset_files);
 
-  // Add GNUPlot instances into HTML
-  HTMLPlot html;
-  auto names = gnuplot.getInstanceNames();
-
-  for (const auto& name : names)
-  {
-    std::string output;
-    gnuplot.getInstanceOutput(name, output);
-    html.writeline(output);
-  }
-
-  html.dump();
+  ROS_WARN("Press Ctl-C to terminate GNUPlot");
+  ros::waitForShutdown();
 
   return 0;
 }
